Fix Version 1.0.13

+++ lib/sow-util.ts
@@ -13,9 +13,10 @@ const _isPlainObject = ( obj: any ): obj is { [x: string]: any; } => {
 const _extend = ( destination: any, source: any ): { [x: string]: any; } => {
     if ( !_isPlainObject( destination ) || !_isPlainObject( source ) )
         throw new TypeError( `Invalid arguments defined. Arguments should be Object instance. destination type ${typeof ( destination )} and source type ${typeof ( source )}` );
+    // tslint:disable-next-line: forin
     for ( const property in source ) {
-        if ( property === "__proto__" || property === "constructor" ) continue;
-        if ( !source.hasOwnProperty( property ) ) continue;
+        // if ( property === "__proto__" || property === "constructor" ) continue;
+        // if ( !source.hasOwnProperty( property ) ) continue;
         destination[property] = source[property];
     }
     return destination;
@@ -26,8 +27,8 @@ const _deepExtend = ( destination: any, source: any ): { [x: string]: any; } => 
         throw new TypeError( `Invalid arguments defined. Arguments should be Object instance. destination type ${typeof ( destination )} and source type ${typeof ( source )}` );
     // tslint:disable-next-line: forin
     for ( const property in source ) {
-        if ( property === "__proto__" || property === "constructor" ) continue;
-        if ( !source.hasOwnProperty( property ) ) continue;
+        // if ( property === "__proto__" || property === "constructor" ) continue;
+        // if ( !source.hasOwnProperty( property ) ) continue;
         const s = source[property];
         const d = destination[property];
         if ( _isPlainObject( d ) && _isPlainObject( s ) ) {
@@ -94,9 +95,16 @@ export namespace Util {
         }
     }
     export function copyFileSync( src: string, dest: string ): void {
-        if ( _fs.existsSync( dest ) ) {
+        let parse = _path.parse( src );
+        if ( !parse.ext )
+            throw new Error( "Source file path required...." );
+        parse = _path.parse( dest );
+        if ( !parse.ext )
+            throw new Error( "Dest file path required...." );
+        if ( !_fs.existsSync( src ) )
+            throw new Error( `Source directory not found ${src}` );
+        if ( _fs.existsSync( dest ) )
             _fs.unlinkSync( dest );
-        }
         _fs.copyFileSync( src, dest );
     }
     export function rmdirSync( path: string ): void {
